+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.4.8
 
+- Support `checkbox` `checked` bidirectional data-binding (with `<=>` sign)
 - Update default dimensions of meta-example to `1000x500`
 - Fix minor issue with meta-example showing basic examples as advanced